requirements.txt
PyQt6>=6.5.0
psutil>=5.9.0
PyYAML>=6.0
pywin32>=305
requests>=2.31.0